Fix build error: wrapped useSearchParams in Suspense
